Move ReadLink from files.cpp to cuttlefish/posix
