+++ base/cvd/cuttlefish/common/libs/utils/BUILD.bazel
@@ -258,6 +258,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/posix:readlink",
         "//cuttlefish/result",
         "//libbase",
         "@abseil-cpp//absl/log",
