+++ base/cvd/cuttlefish/common/libs/utils/proc_file_utils.cpp
@@ -37,6 +37,7 @@
 #include "cuttlefish/common/libs/fs/shared_buf.h"
 #include "cuttlefish/common/libs/fs/shared_fd.h"
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/posix/readlink.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
