+++ base/cvd/cuttlefish/host/libs/command_util/BUILD.bazel
@@ -37,6 +37,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:json",
         "//cuttlefish/host/libs/config:config_instance_derived",
         "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/posix:readlink",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/result",
